% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help_functions.R
\name{comment_end}
\alias{comment_end}
\title{Comment end for R Markdown/HTML}
\usage{
comment_end(html = FALSE)
}
\arguments{
\item{html}{should the comment be used in HTML (FASLE by default)}
}
\description{
Comment end for R Markdown/HTML
}
\keyword{internal}

