% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comment_width.R
\name{comment_width}
\alias{comment_width}
\title{Function to specify the width of the comment}
\usage{
comment_width(width = "option", a4portrait_width = 80)
}
\arguments{
\item{width}{is the prefered width. Could be numeric (number of characters), a paper size
(currently \code{"a4portrait"}, \code{"a4landscape"}, \code{"a3portrait"} or \code{"a3landscape"}),
 \code{"script_width"} (= \code{getOption("width") - 5}) or \code{"option"} (to get data from global option \code{"comment_width"}).
 Default is \code{"option"} but if a global option does not exist, "a4portrait".}

\item{a4portrait_width}{specifies the number of characters that can be printed on a single line on a a4 paper
in portrait orientation. This value is usually 80 (or in the range from 60 to 75). Here a lower value is set by default
due to experimentation on the authors own computer. Please contact the author if this value seems strange.
Note however that it is usully more sufficient to use a global option for the \code{width} parameter than to
manually change this value (whih is however possible for increased flexibility). This value depends on margins and font size
when printing.}
}
\value{
An integer specifiing the text width (in number of characters) to be used in a comment.
}
\description{
Helper function to the commet family of functions. Can take both numerical values as the width
in characters or a prespecified standard through a character string.
}
\details{
The width is specified in order to fit the comment on a printed page or in a script file showed in RStudio
(hence  the -5 in \code{getOption("width") - 5} to adjust for the line numbering in the script file compared to the console.
It is also possible to use a global option \code{comment_width}. This is not a standard option and has to be manually
specified (for example in a .Rprofile-file).
}
\examples{
comment_width()
comment_width(42)
comment_width("a4portrait")

\dontrun{
# We can set a global option for the comment_width
options(comment_width = 80)
header_comment("Test", "A small test")
}

}

