% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comr.R
\name{comr}
\alias{comr}
\title{Comments for R}
\usage{
comr(arg1 = "Headline 1", arg2 = 1)
}
\arguments{
\item{arg1}{String for title}

\item{arg2}{Every object possible}
}
\description{
Gives comments with left indentation. Without capitalization when arg2 is
missing.
}
\examples{
comr("Headline 1")

}
