% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comr.R
\name{comr1}
\alias{comr1}
\title{Centered Comments for R}
\usage{
comr(title = "Headline 1")
}
\arguments{
\item{title}{String for title}

\item{width}{Number of characters (integer) for width of line. Default is 80.}
}
\description{
Gives centered comments. Always capitalized.
}
\examples{
comr("headline")

}
