% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comr.R
\name{comr_h0}
\alias{comr_h0}
\title{Headline 1 comment for R}
\usage{
comr_h1(title = "Headline 1")
}
\arguments{
\item{title}{String for title}

\item{width}{Number of characters (integer) for width of line. Default is 80.}
}
\description{
Gives centered comments. Always capitalized.
}
\examples{
comr_h1("headline")

}
