% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comtex.R
\name{comtex_h2}
\alias{comtex_h2}
\title{Centered Headline 2 for LaTeX}
\usage{
comtex_h2("Headline 1")
}
\arguments{
\item{title}{String for title.}

\item{width}{Number of characters for width of caption}
}
\description{
Gives centered comments for LaTeX. Always capitalized.
}
\examples{
comtex_h2("Headline 1")

}
