% Generated by roxygen2 (4.0.2): do not edit by hand
\name{empty_line}
\alias{empty_line}
\title{Helper function to create string with hashtags and spaces}
\usage{
empty_line(..., token = "#")
}
\arguments{
\item{...}{arguments passed to \code{comment_width}}

\item{token}{a character specifying the comment symbol. "#" by default.}
}
\description{
Helper function to create string with hashtags and spaces
}

