% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/help_functions.R
\name{out}
\alias{out}
\title{Handle output of the comment}
\usage{
out(msg, html, clipboard = TRUE, verbose = TRUE)
}
\arguments{
\item{msg}{the comment to return}

\item{clipboard,verbose,html}{See \code{\link{commentr}}}
}
\description{
Handle output of the comment
}
\keyword{internal}

