% Generated by roxygen2 (4.0.2): do not edit by hand
\name{comment_start}
\alias{comment_start}
\title{Comment start for R Markdown/HTML}
\usage{
comment_start(html = FALSE)
}
\arguments{
\item{html}{should the comment be used in HTML (FASLE by default)}
}
\description{
Comment start for R Markdown/HTML
}

