% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comr.R
\name{comr1}
\alias{comr1}
\title{Centered Comments for R}
\usage{
comr(arg1 = "Headline 1")
}
\arguments{
\item{arg1}{String for title}
}
\description{
Gives centered comments. Always capitalized.
}
\examples{
comr("headline")

}
