% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comtex.R
\name{comtex}
\alias{comtex}
\title{Centered Comments for LaTeX}
\usage{
comr(arg1 = "Headline 1")
}
\arguments{
\item{arg1}{String for title}
}
\description{
Gives centered comments for LaTeX. Always capitalized.
}
\examples{
comr("Headline 1")

}
