% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/help_functions.R
\name{empty_line}
\alias{empty_line}
\title{Helper function to create string with hashtags and spaces}
\usage{
empty_line(..., token = "#")
}
\arguments{
\item{...}{arguments passed to \code{comment_width}}

\item{token}{a character specifying the comment symbol. "#" by default.}
}
\description{
Helper function to create string with hashtags and spaces
}
\keyword{internal}

